backend/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
langchain==0.1.16
langchain_openai==0.1.3
langchain_text_splitters==0.0.1
python-dotenv==1.0.1
